added comments, added stuff to the if statements, and added quotation marks around the insert constants
Added comments explaining what everything does, to the best of my ability. As well as added single quotation marks around each of the phrases in the insertX, insertY, and insertZ arrays. Lastly, i added stuff to set the name to bob if the name box is left empty, added conversions for if the user checks the uk button, and made the story.textContent equal newStory, so that the story changes every time.